Restrict includes to PROJECT/etc/apache/conf.d only
Complicated story behind this, but it's revealed a minor issue with the current set up. We have a set up like this:

PROJECT/etc/apache/conf.d
PROJECT/etc/apache/conf.d/project
PROJECT/etc/apache/conf.d/project_ssl

The idea is that we can have project wide configuration in conf.d and, say, ssl specific bits in project_ssl.

However the configuration looks like:

      ## Load additional static includes
      Include "<PROJECT>/etc/apache/conf.d/"

which includes that directory *and* subdirectories, which defeats the point of the subdirectories.

This uses a wildcard at the top level, which will prevent the inclusion of subdirs.